requirements.txt
# Stratos1 Trading Bot - Dependencies
# Install: pip install -r requirements.txt

# Telegram
telethon>=1.36.0          # User session for reading private groups
python-socks[asyncio]>=2.4.4  # SOCKS proxy support for Telethon

# Bybit
pybit>=5.8.0              # Official Bybit V5 SDK (REST + WebSocket)

# Database
aiosqlite>=0.20.0         # Async SQLite

# Config & Validation
pydantic>=2.6.0           # Config validation
pydantic-settings>=2.2.0  # .env loading
tomli>=2.0.1              # TOML parsing (Python <3.11 compat)
tomli-w>=1.0.0            # TOML writing

# Logging
structlog>=24.1.0         # Structured logging

# Utilities
python-dotenv>=1.0.1      # .env file loading
pytz>=2024.1              # Timezone handling
apscheduler>=3.10.4       # Scheduled tasks (reports)

# Build
pyinstaller>=6.4.0        # Windows executable packaging

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
